feat/refactor: refactored controller to work with new promise based connection and implemented authorization

+++ backend/src/controllers/categories-controller.ts
@@ -1,13 +1,14 @@
 import type { Request, Response } from "express"
 import { connection } from "../utils/db-connection"
-import { handleQueryOutput } from "../utils/query-handling";
+import { handleUser } from "../utils/query-handling";
 
 export const createCategory = async(req: Request, res: Response) => {
-    await connection.execute(
-        `INSERT INTO categories (categoryName) VALUES (?)`,
-        [req.body["categoryName"]],
-        handleQueryOutput(201, res)
-    );
+    const user =  await handleUser(req, res);
+    if (!user) {
+        return;
+    }
+
+    res.sendStatus(201);
 };
 
 export const listCategories = async(req: Request, res: Response) => {
@@ -19,17 +20,27 @@ export const listCategories = async(req: Request, res: Response) => {
 };
 
 export const updateCategory = async(req: Request, res: Response) => {
+    const user = await handleUser(req, res);
+    if (!user) {
+        return;
+    }
+
     await connection.execute(
         `UPDATE categories SET categoryName = ? WHERE categoryId = ?`,
-        [req.body["categoryName"], req.params["id"]],
-        handleQueryOutput(200, res)
+        [req.body["categoryName"], req.params["id"]]
     )
+    res.sendStatus(200)
 };
 
 export const deleteCategory = async(req: Request, res: Response) => {
+    const user = await handleUser(req, res);
+    if (!user) {
+        return;
+    }
+
     await connection.execute(
         `DELETE FROM categories WHERE categoryId = ?`,
-        [req.params["id"]],
-        handleQueryOutput(200, res)
+        [req.params["id"]]
     )
+    res.sendStatus(200);
 };
